+++ unreal/Plugins/RaftSim/Source/RaftSimEditor/Private/Environment/RaftSimEditorEnvironmentInternal.h
@@ -984,6 +984,8 @@ UMaterial* LoadOrCreateLandscapeCandidateSolverSurfaceWaterParent(
 
 UMaterial* LoadOrCreatePacuareRainforestWaterParent(FString& OutSummary);
 
+UMaterial* LoadOrCreateFutaleufuTerminatorWaterParent(FString& OutSummary);
+
 UMaterialParameterCollection* LoadOrCreateRaftFoamOcclusionCollection(
     FString& OutSummary);